Implement performance tests that use CUDA memory
This is useful for PC systems with a dGPU attached. On such a system,
rdma-malloc-*-perf test the performance of FPGA access to the host's
system memory, whereas rdma-cuda-*-perf test the performance of FPGA
access to the dGPU's memory.

On Tegra systems, iGPU and system memory are the same thing, so the
two applications are equivalent, aside from the use of CUDA APIs for
acquiring and pinning the memory prior to FPGA access.